REFACTOR: 상태 및 핸들러 useVote 훅으로 분리
